+++ src/componentes/escolas/Receitas/Formularios/index.js
@@ -11,7 +11,7 @@ import {
     getTabelasReceita,
     getRepasse
 } from '../../../../services/escolas/Receitas.service';
-import {round, trataNumericos, periodoFechado} from "../../../../utils/ValidacoesAdicionaisFormularios";
+import {round, trataNumericos, periodoFechado, comparaObjetos} from "../../../../utils/ValidacoesAdicionaisFormularios";
 import {ReceitaSchema} from '../Schemas';
 import moment from "moment";
 import {useParams} from 'react-router-dom';
@@ -79,7 +79,6 @@ export const ReceitaForm = props => {
             if (uuid) {
                 getReceita(uuid).then(response => {
                     const resp = response.data;
-                    console.log("RESPONSE ", resp);
                     const init = {
                         tipo_receita: resp.tipo_receita.id,
                         detalhe_tipo_receita: resp.detalhe_tipo_receita,
@@ -116,20 +115,6 @@ export const ReceitaForm = props => {
     }, [tabelas, initialValue.tipo_receita]);
 
     const servicoDeVerificacoes = (e, values, errors) =>{
-
-        //console.log("Objeto ", objetoParaComparacao);
-        //console.log("Values ", values);
-
-        let sao_iguais= comparaObjetos(values, objetoParaComparacao)
-
-        console.log("sao_iguais  ", sao_iguais);
-
-        if (values === objetoParaComparacao){
-            console.log("OBJETOS IGUAIS ")
-        }else {
-            console.log("OBJETOS DIFERENTES ")
-        }
-
         // Validando se receita é conferida
         if (Object.entries(errors).length === 0 ) {
             if (values.conferido) {
@@ -138,37 +123,6 @@ export const ReceitaForm = props => {
             }
         }
     };
-
-    const  comparaObjetos = (objetoA, objetoB) =>{
-
-        //Busca as chaves do objetoA e objetoB
-        //utilizando o "let" o escopo da variável é limitado para o bloco.
-        //Object.keys retornará um array com todas as chaves do objeto.
-        let aChaves = Object.keys(objetoA),
-            bChaves = Object.keys(objetoB);
-
-        //Compara os tamanhos, se forem diferentes retorna falso pois
-        //o numero de propriedades é diferente, logo os objetos são diferentes
-        if (aChaves.length !== bChaves.length) {
-            return false;
-        }
-
-        //Verifico se existe algum elemento com valor diferente nos objetos.
-        //o array.some executa uma função(passada por parâmetro) para cada valor
-        //do array. Essa função deve executar um teste, se para algum dos valores
-        //o teste é verdadeiro, a execução é interrompida e true é retornado.
-        //Do contrário, se o teste nunca for verdadeiro ele retornará false
-        //após executar o teste para todos valores do array.
-        //Estou basicamente verficando se existe diferença entre dois valores do objeto.
-
-        let saoDiferentes = aChaves.some((chave) => {
-            return objetoA[chave] !== objetoB[chave];
-        });
-
-        //como saoDiferentes contém true caso os objetos sejam diferentes eu
-        //simplesmente nego esse valor para retornar que os objetos são iguais (ou não).
-        return !saoDiferentes;
-    }
 
     const onSubmit = async (values) => {
         // Validando e ou removendo e_devolucao
@@ -733,7 +687,7 @@ export const ReceitaForm = props => {
 
                             {/*Botões*/}
                             <div className="d-flex justify-content-end pb-3" style={{marginTop: '60px'}}>
-                                <button type="reset" onClick={values === objetoParaComparacao ? onCancelarTrue : onShowModal} className="btn btn btn-outline-success mt-2 mr-2">Voltar
+                                <button type="reset" onClick={comparaObjetos(values,objetoParaComparacao) ? onCancelarTrue : onShowModal} className="btn btn btn-outline-success mt-2 mr-2">Voltar
                                 </button>
                                 {uuid ?
                                     <button disabled={readOnlyBtnAcao} type="reset" onClick={onShowDeleteModal} className="btn btn btn-danger mt-2 mr-2">Deletar</button> : null
